moved dark mode logic to a custom hook
