+++ src/popup/App.jsx
@@ -1,12 +1,9 @@
 import { Space, Divider, ConfigProvider, theme, Layout } from 'antd';
-import React, { useState, useEffect } from 'react';
-
-// Utils
-import Logger from '../scripts/utils/logger.js';
 
 // Hooks
 import { useExtensionSettings } from './hooks/useExtensionSettings';
 import { useScreenshotCapture } from './hooks/useScreenshotCapture.js';
+import { useDarkMode } from './hooks/useDarkMode.js';
 
 // Components
 import Header from './components/Header.jsx';
@@ -47,56 +44,16 @@ export default function App() {
     handleCaptureScreenshot,
   } = useScreenshotCapture(isRestricted);
 
-  const [darkMode, setDarkMode] = useState(false);
-
-  /**
-   * Toggles dark mode on or off.
-   *
-   * @param {boolean} checked - True if dark mode is enabled, false otherwise.
-   */
-  const handleToggleDarkMode = async checked => {
-    setDarkMode(checked);
-    // Save the dark mode preference to local storage
-    try {
-      await chrome.storage.local.set({ darkMode: checked });
-      Logger.info(`Dark mode preference saved: ${checked}`);
-    } catch (error) {
-      Logger.error('Error saving dark mode preference:', error);
-    }
-  };
-
-  // Loads the dark mode preference from local storage when the component mounts.
-  useEffect(() => {
-    const loadDarkModePreference = async () => {
-      try {
-        const { darkMode: savedDarkMode } = await chrome.storage.local.get(
-          'darkMode'
-        );
-
-        // If a preference is found, set it; otherwise, default to light mode
-        if (savedDarkMode !== undefined) {
-          handleToggleDarkMode(savedDarkMode);
-        } else {
-          Logger.warn(
-            'No dark mode preference found, defaulting to light mode.'
-          );
-          handleToggleDarkMode(false);
-        }
-      } catch (error) {
-        Logger.error('Error loading dark mode preference:', error);
-      }
-    };
-    loadDarkModePreference();
-  }, []);
+  const { isDarkMode, handleToggleDarkMode } = useDarkMode();
 
   return (
     <ConfigProvider
       theme={{
-        algorithm: darkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
+        algorithm: isDarkMode ? theme.darkAlgorithm : theme.defaultAlgorithm,
         token: {
           colorPrimary: '#2374ab', // Ant Design primary color (Border Patrol blue)
-          colorBgContainer: darkMode ? '#141414' : '#ffffff', // Background color for containers
-          colorText: darkMode ? '#ffffff' : '#132a3e', // Text color
+          colorBgContainer: isDarkMode ? '#141414' : '#ffffff', // Background color for containers
+          colorText: isDarkMode ? '#ffffff' : '#132a3e', // Text color
         },
         components: {
           Layout: {
@@ -111,7 +68,7 @@ export default function App() {
         <FeatureToggle
           label='Dark Mode'
           id='dark-mode-toggle'
-          checked={darkMode}
+          checked={isDarkMode}
           onChange={handleToggleDarkMode}
           ariaLabel='Toggle dark mode'
         />
